Change tableLen type from int to size_t

+++ src/m1_encoder.h
@@ -3,13 +3,12 @@
 
 #include "../include/types.h"
 #include <stddef.h>
-
 int loadMorseTable(AlphabetSet *outAlphabet, size_t maxEntries, const char *dataPath);
 int encodeText(const char *input, const AlphabetSet *alphabet, char *outBuf, size_t outBufSize);
 int decodeMorse(const char *morseInput, const AlphabetSet *alphabet, char *outBuf, size_t outBufSize);
 int validateMorseInput(const char *morseInput);
-int lookupCharacter(const MorseEntry *table, int tableLen, char ch);
+int lookupCharacter(const MorseEntry *table, size_t tableLen, char ch);
 const char *getMorseForChar(const AlphabetSet *alphabet, char ch);
 int getCharFromMorse(const AlphabetSet *alphabet, const char *morse, char *outChar);
 
-#endif // M1_ENCODER_H
+#endif
